Wifi: Add support for repeater mode
Repeater Mode is a feature where STA network is shared to
peers devices via hotspot. As part of this implementation
introduce a variable in WifiConfiguration to check whether
connected AP will share the credentials with hotspot or not.
If this value is set to true then hotspot will turn on using
the ssid and passphrase of the connected AP.

Change-Id: I28d0a60b240f29ca11b831f069fe61e3b3d55a6f
CRs-Fixed: 2217486
Signed-off-by: DennySPB <dennyspb@gmail.com>
Signed-off-by: Anirudh Gupta <anirudhgupta109@gmail.com>